refactor to have only onRollDamage function
